refactor(frame-aggregate): move aggregate helper back to mloda-testing
The earlier rename to mloda.community.feature_groups.data_operations.aggregate_helpers
was justified by the pandas month/year per-row relativedelta loop needing the helper
at runtime in the community wheel. That loop was deleted in 790e76a; pandas/sqlite
now reject month/year at match time and no community runtime code imports the helper.
Its only consumers are the testing reference implementation and one test, so the
helper moves back to mloda.testing.feature_groups.data_operations.aggregation_helpers.
