+++ mloda/testing/feature_groups/data_operations/aggregation_helpers.py
@@ -1,10 +1,8 @@
-"""Shared pure-Python aggregation helpers.
+"""Shared aggregation helpers for test reference implementations.
 
-Used by the reference implementation and by per-row community implementations
-that need calendar-anchored windowing. The helper functions below compute
-scalar aggregates over plain Python lists. They are used by frame_aggregate
-(per-window aggregation) and as a fallback for operations that PyArrow's
-native group_by API does not support directly.
+The helper functions below compute scalar aggregates over plain Python lists.
+They are used by frame_aggregate (per-window aggregation) and as a fallback
+for operations that PyArrow's native group_by API does not support directly.
 """
 
 from __future__ import annotations
